requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
pytesseract>=0.3.10
easyocr>=1.7.0
opencv-python>=4.8.0
pillow>=10.0.0
requests>=2.31.0
geopandas>=0.14.0
shapely>=2.0.0
pandas>=2.0.0
plotly>=5.17.0
python-dotenv>=1.0.0
numpy>=1.24.0
fiona>=1.9.0
pyproj>=3.6.0
